requirements.txt
Django>=2.2,<2.3
wagtail>=2.7,<2.8
python-dotenv
django_twilio
